feat(battlefield): add verdant_battlefield.glb and rename ResourceMap to Battlefield
- Copy verdant_battlefield.glb (29 MB) to apps/web/public/assets/models/
- Rename game/ResourceMap/ → game/Battlefield/ to better reflect its purpose
- Rename ResourceMapPage.css → BattlefieldPage.css
- Update all imports in UnifiedMapLayers, UnifiedMapScene, FarmingPage
